feat: Pure Builder
This commit removes the installation of NGINX and only focuses on the assemble step, which builds the Web application.

The output of running 'npm build' is now copied to '/opt/app-root/output'

The run phase will now just use the Serve module to serve the static content. This is **not** recommended

This image should just be used as a builder image, and should be part of an Openshift Chain Build https://docs.openshift.org/3.9/dev_guide/builds/advanced_build_operations.html\#dev-guide-chaining-builds