feat: Optimizations for a better developer workflow
This includes adding a NPM_RUN_COMMAND variable that allows the user to specify what to run in the run phase(when not using a chain build).  For example, you could run a react dev server that could watch for changes in your code.

Also fixes an issue where the DEPLOY_DIR wasn't set correctly and the app would show a directory listing instead.

Adds a DEPLOY_PORT variable to override the port that the serve module uses.

fixes #8